requirements.txt
Flask==1.0.2
Flask-login
flask-sqlalchemy
rauth
flask-wtf
python-dotenv
psycopg2
flask-script
flask-migrate
